requirements.txt
solidpython
matplotlib
numpy
